Move post processing from Opt to seperate file
